feat(volunteer): create new file for getVolunteer utils funcs called volunteerLogic
